requirements.txt
# Web Framework
Flask==3.0.3
gunicorn==23.0.0
Werkzeug==3.1.3

# Authentication
flask-login==0.6.3
bcrypt==4.2.0

# Google APIs for YouTube OAuth
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.147.0

# Groq AI API
groq==0.11.0
httpx==0.27.0

# Video & Image Processing
opencv-python-headless==4.10.0.84
ffmpeg-python==0.2.0
Pillow==11.0.0
numpy==2.1.2

# YouTube Downloader
yt-dlp==2024.11.4

# HTTP & Environment
requests==2.32.3
urllib3==2.2.3
python-dotenv==1.0.1

# Session Management
Flask-Session==0.8.0

# Payments (optional — app works without it)
stripe==7.0.0

# Build Tools
setuptools>=75.1.0
wheel>=0.44.0
